planSidewinderTour update
gamma is no longer needed as an input of the function. Instead, previous tour (tour0) and grid (grid0) are. With these variables, we can determine in which direction is the coverage path going and, besides, determine the next planned observation (according to the previous tour). replanningSidewinder and sidewinder have been modified accordingly to this new implementation.